# Refactor validation: resolve fragments once