--- a/lib/graphql/static_validation/rules/fragments_are_finite.rb
+++ b/lib/graphql/static_validation/rules/fragments_are_finite.rb
@@ -5,38 +5,13 @@ module GraphQL
       include GraphQL::StaticValidation::Message::MessageHelper
 
       def validate(context)
-        context.visitor[GraphQL::Language::Nodes::FragmentDefinition] << ->(node, parent) {
-          if has_nested_spread?(node, [], context)
-            context.errors << message("Fragment #{node.name} contains an infinite loop", node)
+        context.visitor[GraphQL::Language::Nodes::Document].leave << ->(_n, _p) do
+          dependency_map = context.dependencies
+          dependency_map.cyclical_definitions.each do |defn|
+            if defn.is_a?(GraphQL::Language::Nodes::FragmentDefinition)
+              context.errors << message("Fragment #{defn.name} contains an infinite loop", defn)
+            end
           end
-        }
-      end
-
-      private
-
-      def has_nested_spread?(fragment_def, parent_fragment_names, context)
-        nested_spreads = get_spreads(fragment_def.selections)
-
-        nested_spreads.any? do |spread|
-          nested_def = context.fragments[spread.name]
-          if nested_def.nil?
-            # this is another kind of error
-            false
-          else
-            parent_fragment_names.include?(spread.name) || has_nested_spread?(nested_def, parent_fragment_names + [fragment_def.name], context)
-          end
-        end
-      end
-
-      # Find spreads contained in this selection & return them in a flat array
-      def get_spreads(selection)
-        case selection
-        when GraphQL::Language::Nodes::FragmentSpread
-          [selection]
-        when GraphQL::Language::Nodes::Field, GraphQL::Language::Nodes::InlineFragment
-          get_spreads(selection.selections)
-        when Array
-          selection.map { |s| get_spreads(s) }.flatten
         end
       end
     end