--- a/lib/graphql/static_validation/rules/fragments_are_used.rb
+++ b/lib/graphql/static_validation/rules/fragments_are_used.rb
@@ -5,48 +5,19 @@ module GraphQL
       include GraphQL::StaticValidation::Message::MessageHelper
 
       def validate(context)
-        v = context.visitor
-        used_fragments = []
-        defined_fragments = []
-
-        v[GraphQL::Language::Nodes::Document] << ->(node, parent) {
-          defined_fragments = node.definitions
-            .select { |defn| defn.is_a?(GraphQL::Language::Nodes::FragmentDefinition) }
-            .map { |node| FragmentInstance.new(node: node) }
-        }
-
-        v[GraphQL::Language::Nodes::FragmentSpread] << ->(node, parent) {
-          used_fragments << FragmentInstance.new(node: node)
-          if defined_fragments.none? { |defn| defn.name == node.name }
-            GraphQL::Language::Visitor::SKIP
+        context.visitor[GraphQL::Language::Nodes::Document].leave << ->(_n, _p) do
+          dependency_map = context.dependencies
+          dependency_map.unmet_dependencies.each do |op_defn, spreads|
+            spreads.each do |fragment_spread|
+              context.errors << message("Fragment #{fragment_spread.name} was used, but not defined", fragment_spread)
+            end
           end
-        }
-        v[GraphQL::Language::Nodes::Document].leave << ->(node, parent) { add_errors(context, used_fragments, defined_fragments) }
-      end
 
-      private
-
-      def add_errors(context, used_fragments, defined_fragments)
-        undefined_fragments = find_difference(used_fragments, defined_fragments.map(&:name))
-        undefined_fragments.each do |fragment|
-          context.errors << message("Fragment #{fragment.name} was used, but not defined", fragment.node)
-        end
-
-        unused_fragments = find_difference(defined_fragments, used_fragments.map(&:name))
-        unused_fragments.each do |fragment|
-          context.errors << message("Fragment #{fragment.name} was defined, but not used", fragment.node)
-        end
-      end
-
-      def find_difference(fragments, allowed_fragment_names)
-        fragments.select {|f| f.name && !allowed_fragment_names.include?(f.name) }
-      end
-
-      class FragmentInstance
-        attr_reader :name, :node
-        def initialize(node:)
-          @node = node
-          @name = node.name
+          dependency_map.unused_dependencies.each do |fragment|
+            if !fragment.name.nil?
+              context.errors << message("Fragment #{fragment.name} was defined, but not used", fragment)
+            end
+          end
         end
       end
     end