--- a/spec/support/dummy/schema.rb
+++ b/spec/support/dummy/schema.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+require_relative "./data"
 module Dummy
   class NoSuchDairyError < StandardError; end