This is a long-standing pet peeve of mine: several of the validators make their own ad-hoc traversals of the AST, which defeats the purpose of the visitor and wastes time. 

I've extracted some code from #268 and applied it here. I think there might be a couple other places I can use this, I'll take another look before I merge.

So far, we get a slight speed-up: 

- Master: 

  ```
  validate    285.846  (± 4.5%) i/s -      1.428k in   5.006162s
  validate    284.037  (± 3.9%) i/s -      1.428k in   5.035883s
  validate    273.442  (± 7.3%) i/s -      1.372k in   5.049888s
  validate    279.002  (± 3.9%) i/s -      1.404k in   5.040443s
  ```

- This branch: 

  ```
  validate    294.472  (± 5.1%) i/s -      1.479k in   5.036770s
  validate    297.198  (± 4.4%) i/s -      1.500k in   5.057434s
  validate    295.390  (± 3.4%) i/s -      1.479k in   5.012943s
  ```