@rmosolgo 2017-01-24
Done better in #504